code-review: read the generated IR through one helper, and lend the lifetime markers a shared receiver
